# ENH: Faster adapter loading if there are a lot of target modules