This is an optimization to reduce the number of entries in the `target_modules` set. The reason is that in some circumstances, `target_modules` can contain hundreds of entries. Since each target module is checked against each module of the net (which can be thousands), this can become quite expensive when many adapters are being added. Often, the `target_modules` can be condensed in such a case, which speeds up the process.

A context in which this can happen is when diffusers loads non-PEFT LoRAs. As there is no meta info on target_modules in that case, they are just inferred by listing all keys from the state_dict, which can be quite a lot. See: https://github.com/huggingface/diffusers/issues/9297

As there is a small chance for undiscovered bugs, we apply this optimization only if the list of `target_modules` is sufficiently big. Therefore, for normal PEFT users, this should not have any effect.

Example:

```python
>>> from peft.tuners.tuners_utils import _find_minimal_target_modules
>>> target_modules = [f"model.decoder.layers.{i}.self_attn.q_proj" for i in range(100)]
>>> target_modules += [f"model.decoder.layers.{i}.self_attn.v_proj" for i in range(100)]
>>> other_module_names = [f"model.encoder.layers.{i}.self_attn.k_proj" for i in range(100)]
>>> _find_minimal_target_modules(target_modules, other_module_names)
{"q_proj", "v_proj"}
```

As shown in https://github.com/huggingface/diffusers/issues/9297, the speed improvements for loading many diffusers LoRAs can be substantial. When loading 30 adapters, the time would go up from 0.6 sec per adapter to 3 sec per adapter. With this fix, the time goes up from 0.6 sec per adapter to 1 sec per adapter.